Remove settings tree from published types for now
