+++ packages/studio-base/src/components/SettingsTreeEditor/FieldEditor.tsx
@@ -18,13 +18,13 @@ import {
 import { useMemo } from "react";
 import { DeepReadonly } from "ts-essentials";
 
-import { SettingsTreeField } from "@foxglove/studio";
 import MessagePathInput from "@foxglove/studio-base/components/MessagePathSyntax/MessagePathInput";
 import Stack from "@foxglove/studio-base/components/Stack";
 
 import { ColorPickerInput } from "./inputs/ColorPickerInput";
 import { ColorScalePicker } from "./inputs/ColorScalePicker";
 import { NumberInput } from "./inputs/NumberInput";
+import { SettingsTreeField } from "./types";
 
 const StyledToggleButtonGroup = muiStyled(ToggleButtonGroup)(({ theme }) => ({
   backgroundColor: theme.palette.action.hover,
