+++ packages/studio-base/src/components/SettingsTreeEditor/index.tsx
@@ -8,10 +8,10 @@ import { AppBar, IconButton, TextField, styled as muiStyled, List } from "@mui/m
 import { useCallback, useState } from "react";
 import { DeepReadonly } from "ts-essentials";
 
-import { SettingsTree } from "@foxglove/studio";
 import Stack from "@foxglove/studio-base/components/Stack";
 
 import { NodeEditor } from "./NodeEditor";
+import { SettingsTree } from "./types";
 
 const StyledAppBar = muiStyled(AppBar, { skipSx: true })(({ theme }) => ({
   top: -1,
