+++ packages/studio-base/src/panels/Teleop/TeleopPanel.tsx
@@ -4,16 +4,16 @@
 
 import { Stack, Typography } from "@mui/material";
 import { set } from "lodash";
-import { ReactNode, useCallback, useEffect, useLayoutEffect, useState } from "react";
+import { ReactNode, useCallback, useContext, useEffect, useLayoutEffect, useState } from "react";
 import { DeepPartial } from "ts-essentials";
 
 import { definitions as commonDefs } from "@foxglove/rosmsg-msgs-common";
+import { PanelExtensionContext, Topic } from "@foxglove/studio";
 import {
-  PanelExtensionContext,
   SettingsTreeAction,
   SettingsTreeNode,
-  Topic,
-} from "@foxglove/studio";
+} from "@foxglove/studio-base/components/SettingsTreeEditor/types";
+import { PanelSettingsEditorContext } from "@foxglove/studio-base/context/PanelSettingsEditorContext";
 import ThemeProvider from "@foxglove/studio-base/theme/ThemeProvider";
 
 import DirectionalPad, { DirectionalPadAction } from "./DirectionalPad";
@@ -131,6 +131,8 @@ function TeleopPanel(props: TeleopPanelProps): JSX.Element {
   const [currentAction, setCurrentAction] = useState<DirectionalPadAction | undefined>();
   const [topics, setTopics] = useState<readonly Topic[]>([]);
 
+  const { publishPanelSettingsTree } = useContext(PanelSettingsEditorContext);
+
   // resolve an initial config which may have some missing fields into a full config
   const [config, setConfig] = useState<Config>(() => {
     const partialConfig = context.initialState as DeepPartial<Config>;
@@ -176,13 +178,17 @@ function TeleopPanel(props: TeleopPanelProps): JSX.Element {
         setColorScheme(renderState.colorScheme);
       }
     };
-  }, [context, settingsActionHandler]);
+  }, [context]);
 
   useEffect(() => {
     const tree = buildSettingsTree(config, topics);
-    context.publishPanelSettingsTree({ settings: tree, actionHandler: settingsActionHandler });
+    // eslint-disable-next-line no-underscore-dangle, @typescript-eslint/no-explicit-any
+    (context as unknown as any).__publishPanelSettingsTree({
+      settings: tree,
+      actionHandler: settingsActionHandler,
+    });
     saveState(config);
-  }, [config, context, saveState, settingsActionHandler, topics]);
+  }, [config, context, publishPanelSettingsTree, saveState, settingsActionHandler, topics]);
 
   // advertise topic
   const { topic: currentTopic } = config;
